Add files via upload
Version 50 

XOR shift churns 64-81 times for seed. 

This seed is a replacement for the fact that ARM does not provide a time seed ( insecure anyway) or a get time assembly call ( lucky as it's insecure anyway ). So we run a XOR shift with a 64 bit variable, churn it 64-81 times. 

Note, the original haveged use of the 32-bit variable may have been insecure anyway. 

Online tests ensure ( "tbc", NOT cpu intensive default ) that there is no leakage of entropy quality.